Improve a Django Project Beta V1.0
The site is up and running as intented. The Menu model has been convered to use a DateField for the expiration date.  Views.py has been updated corespondingly. The SelectDateWidget is up and running properly. The database is optimally queried. More tests need to be written, and the project needs just a little more polishing.